Add ai_utils patch

+++ game/server/ai_utils.h
@@ -70,7 +70,7 @@ public:
 
 	bool IsMarkSet()
 	{
-		return ( m_flMarkTolerance != NO_MARK );
+		return ( m_flMarkTolerance != (float)NO_MARK );
 	}
 
 	bool TargetMoved( CBaseEntity *pEntity )
